+++ code/planecalib/CameraModelCeres.h
@@ -12,18 +12,18 @@ template<class TDistortionModel>
 void CameraModel_<TDistortionModel>::projectFromWorldJacobian(const Eigen::Vector3f &xc, Eigen::Vector3f &ujac, Eigen::Vector3f &vjac) const
 {
 	typedef ceres::Jet<double, 3> TJet;
-	Eigen::Matrix<TJet, 3, 1> x(TJet(xc[0], 0), TJet(xc[1], 1), TJet(xc[2], 2));
+  Eigen::Matrix<TJet, 3, 1> x(TJet(static_cast<double>(xc[0]) , 0), TJet(static_cast<double>(xc[1]), 1),
+      TJet(static_cast<double>(xc[2]), 2));
 
 	Eigen::Matrix<TJet, 2, 1> p;
 
 	projectFromWorld(x,p);
-	ujac[0] = (float)p[0].v[0];
-	ujac[1] = (float)p[0].v[1];
-	ujac[2] = (float)p[0].v[2];
-
-	vjac[0] = (float)p[1].v[0];
-	vjac[1] = (float)p[1].v[1];
-	vjac[2] = (float)p[1].v[2];
+  ujac[0] = static_cast<float>(p[0].v[0]);
+  ujac[1] = static_cast<float>(p[0].v[1]) ;
+  ujac[2] = static_cast<float>(p[0].v[2]);
+  vjac[0] = static_cast<float>(p[1].v[0]);
+  vjac[1] = static_cast<float>(p[1].v[1]);
+  vjac[2] = static_cast<float>(p[1].v[2]);
 }
 
 }
